% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aristolochia_data.R
\docType{data}
\name{aristolochia}
\alias{aristolochia}
\title{Dataset: Germination of seeds of \emph{Aristolochia} sp. as a function of temperature.}
\format{
data.frame containing data set
  \describe{
  \item{\code{trat}}{numeric vector with factor 1}
  \item{\code{resp}}{Numeric vector with response}
  }
}
\usage{
data("aristolochia")
}
\description{
The data come from an experiment conducted at the Seed Analysis
Laboratory of the Agricultural Sciences Center of the State
University of Londrina, in which five temperatures (15, 20, 25,
30 and 35C) were evaluated in the germination of \emph{Aristolochia elegans}.
The experiment was conducted in a completely randomized
design with four replications of 25 seeds each.
}
\examples{
data(aristolochia)
}
\seealso{
\link{cloro}, \link{laranja}, \link{enxofre}, \link{laranja}, \link{mirtilo}, \link{passiflora}, \link{phao}, \link{porco}, \link{pomegranate}, \link{simulate1}, \link{simulate2}, \link{simulate3}, \link{tomate}, \link{weather}
}
\keyword{datasets}
