% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bargraph_function.R
\name{bar_graph}
\alias{bar_graph}
\title{Graph: Bar graph for one factor}
\usage{
bar_graph(model, fill = "lightblue", horiz = TRUE)
}
\arguments{
\item{model}{DIC, DBC or DQL object}

\item{fill}{fill bars}

\item{horiz}{Horizontal Column (\emph{default} is TRUE)}
}
\value{
Returns a bar chart for one factor
}
\description{
This is a function of the bar graph for one factor
}
\examples{
data("laranja")
a=with(laranja, DBC(trat, bloco, resp,
     mcomp = "sk",angle=45,
     ylab = "Number of fruits/plants"))
bar_graph(a,horiz = FALSE)
}
\seealso{
\link{radargraph}, \link{barplot_positive}, \link{plot_TH}, \link{plot_TH1}, \link{corgraph}, \link{spider_graph}, \link{line_plot}, \link{plot_cor}, \link{plot_interaction}, \link{plot_jitter}, \link{seg_graph}, \link{TBARPLOT.reverse}
}
\author{
Gabriel Danilo Shimizu, \email{shimizu@uel.br}

Leandro Simoes Azeredo Goncalves

Rodrigo Yudi Palhaci Marubayashi
}
