% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bean_dataset.R
\docType{data}
\name{bean}
\alias{bean}
\title{Dataset: Bean}
\format{
data.frame containing data set
  \describe{
  \item{\code{trat}}{numeric vector with treatment}
  \item{\code{prod}}{Numeric vector with grain production per plant}
  }
}
\usage{
data("bean")
}
\description{
An experiment to evaluate the effect of different strains of Azospirillum on common bean cultivar
IPR Sabia was carried out in a greenhouse. A completely randomized design with five strains was used.
of Azospirillum (treatments) and five repetitions. The response variable analyzed was grain production per plant
(g plant-1).
}
\examples{
data(bean)
}
\seealso{
\link{aristolochia}, \link{cloro}, \link{laranja}, \link{enxofre}, \link{laranja}, \link{mirtilo}, \link{passiflora}, \link{phao}, \link{porco}, \link{pomegranate}, \link{simulate1}, \link{simulate2}, \link{simulate3}, \link{tomate}, \link{weather}
}
\keyword{datasets}
