% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloro.R
\docType{data}
\name{cloro}
\alias{cloro}
\title{Dataset: Sodium dichloroisocyanurate in soybean}
\format{
data.frame containing data set
  \describe{
  \item{\code{f1}}{Categorical vector with factor 1}
  \item{\code{f2}}{Categorical vector with factor 2}
  \item{\code{bloco}}{Categorical vector with block}
  \item{\code{resp}}{Numeric vector with number nodules}
  }
}
\usage{
data(cloro)
}
\description{
An experiment was conducted in a greenhouse in pots at the State
University of Londrina. The work has the objective of evaluating
the application of sodium dichloroisocyanurate (DUP) in soybean in
4 periods of application in soybean inoculated or not with Rhizobium
and its influence on the number of nodules. The experiment was
conducted in a completely randomized design with five replications.
}
\examples{
data(cloro)
}
\references{
Rony Kauling Tonelli. Efeito do uso de dicloroisocianurato de sodio sobre a nodulacao em raizes de soja. 2016. Trabalho de Conclusao de Curso. (Graduacao em Agronomia) - Universidade Estadual de Londrina.
}
\seealso{
\link{enxofre}, \link{laranja}, \link{mirtilo}, \link{pomegranate}, \link{porco}, \link{sensorial}, \link{simulate1}, \link{simulate2}, \link{simulate3}, \link{tomate}, \link{weather}, \link{phao}, \link{passiflora}, \link{aristolochia}
}
\keyword{datasets}
