% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coric_function.R
\name{cor_ic}
\alias{cor_ic}
\title{Graph: Plot Pearson correlation with interval of confidence}
\usage{
cor_ic(
  data,
  background = TRUE,
  axis.size = 12,
  ylab = "",
  xlab = "Correlation (r)",
  theme = theme_classic()
)
}
\arguments{
\item{data}{data.frame with responses}

\item{background}{background fill (\emph{default} is TRUE)}

\item{axis.size}{Axes font size (\emph{default} is 12)}

\item{ylab}{Variable response name (Accepts the \emph{expression}() function)}

\item{xlab}{Treatments name (Accepts the \emph{expression}() function)}

\item{theme}{ggplot theme (\emph{default} is theme_classic())}
}
\value{
The function returns a new graphical approach to correlation.
}
\description{
Graph: Plot Pearson correlation with interval of confidence
}
\examples{
data("pomegranate")
cor_ic(pomegranate[,-1])
}
\author{
Gabriel Danilo Shimizu, \email{shimizu@uel.br}

Leandro Simoes Azeredo Goncalves

Rodrigo Yudi Palhaci Marubayashi
}
