% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covercrops_dataset.R
\docType{data}
\name{covercrops}
\alias{covercrops}
\title{Dataset: Covercrops}
\format{
data.frame containing data set
  \describe{
  \item{\code{A}}{Categorical vector with cultivars}
  \item{\code{B}}{Categorical vector with green manure}
  \item{\code{Bloco}}{Categorical vector with block}
  \item{\code{Resp}}{Numeric vector with yield}
  }
}
\usage{
data(covercrops)
}
\description{
Consider a 3 ×3 factorial experiment in randomized blocks, with
4 replications, on the influence of three new soybean cultivars (A1, A2 and A3)
and the use of three types of green manure (B1, B2 and B3) on yield in 100 m2 plots.
}
\examples{
data(covercrops)
}
\seealso{
\link{enxofre}, \link{laranja}, \link{mirtilo}, \link{pomegranate}, \link{porco}, \link{sensorial}, \link{simulate1}, \link{simulate2}, \link{simulate3}, \link{tomate}, \link{weather}, \link{phao}, \link{passiflora}, \link{aristolochia}
}
\keyword{datasets}
