% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emerg.R
\docType{data}
\name{emerg}
\alias{emerg}
\title{Dataset: Emergence of passion fruit seeds over time .}
\format{
data.frame containing data set
  \describe{
  \item{\code{time}}{numeric vector with time}
  \item{\code{resp}}{Numeric vector with emergence}
  }
}
\usage{
data("emerg")
}
\description{
The data come from an experiment conducted at the State University
of Londrina, aiming to study the emergence of yellow passion fruit
seeds over time. Data are partial from one of the treatments studied.
Four replicates with eight seeds each were used.
}
\examples{
data(emerg)
}
\seealso{
\link{aristolochia}, \link{cloro}, \link{laranja}, \link{enxofre}, \link{laranja}, \link{mirtilo}, \link{passiflora}, \link{phao}, \link{porco}, \link{pomegranate}, \link{simulate1}, \link{simulate2}, \link{simulate3}, \link{tomate}, \link{weather}
}
\keyword{datasets}
