% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enxofre.R
\docType{data}
\name{enxofre}
\alias{enxofre}
\title{Dataset: Sulfur data}
\format{
data.frame containing data set
  \describe{
  \item{\code{f1}}{Categorical vector with factor 1}
  \item{\code{f2}}{Categorical vector with factor 2}
  \item{\code{f2}}{Categorical vector with factor 3}
  \item{\code{bloco}}{Categorical vector with block}
  \item{\code{resp}}{Numeric vector}
  }
}
\usage{
data(enxofre)
}
\description{
The experiment was carried out in a randomized block design in a
3 x 3 x 3 triple factorial scheme: syrup volume (75, 225 and 675 L),
sulfur doses (150, 450, 1350) and time of application (vegetative,
complete cycle and reproductive system) with four repetitions.
Yield in kg / ha of soybean was evaluated.
}
\examples{
data(enxofre)
}
\seealso{
\link{cloro}, \link{laranja}, \link{mirtilo}, \link{pomegranate}, \link{porco}, \link{sensorial}, \link{simulate1}, \link{simulate2}, \link{simulate3}, \link{tomate}, \link{weather}, \link{phao}, \link{passiflora}, \link{aristolochia}
}
\keyword{datasets}
