% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eucalyptus_data.R
\docType{data}
\name{eucalyptus}
\alias{eucalyptus}
\title{Dataset: Eucaliptus grandis Barbin (2013)}
\format{
data.frame containing data set
  \describe{
  \item{\code{trati}}{Categorical vector with treatments}
  \item{\code{bloc}}{Categorical vector with block}
  \item{\code{exp}}{Categorical vector with experiment}
  \item{\code{resp}}{Numeric vector}
  }
}
\usage{
data("eucalyptus")
}
\description{
The data refer to the height in meters of *Eucalyptus grandis* plants,
with 7 years of age, from three trials (Araraquara - Exp 1; Bento
Quintino - Exp 2; Mogi-Guacu - Exp 3) in randomized blocks, under
6 progenies. The data were taken from the book by Decio Barbin
(2013) and are from the Instituto Florestal de Tupi/SP.
}
\examples{
data(eucalyptus)
}
\references{
Planejamento e Analise Estatistica de Experimentos Agronomicos (2013) - Decio Barbin - pg. 177
}
\seealso{
\link{cloro}, \link{enxofre}, \link{laranja}, \link{pomegranate}, \link{porco}, \link{sensorial}, \link{simulate1}, \link{simulate2}, \link{simulate3}, \link{tomate}, \link{weather}
}
\keyword{datasets}
