% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAT2DIC_function.R
\name{FAT2DIC}
\alias{FAT2DIC}
\title{Analysis: DIC experiments in double factorial}
\usage{
FAT2DIC(
  f1,
  f2,
  response,
  norm = "sw",
  homog = "bt",
  mcomp = "tukey",
  alpha.f = 0.05,
  alpha.t = 0.05,
  quali = c(TRUE, TRUE),
  grau = NA,
  transf = 1,
  geom = "bar",
  theme = theme_classic(),
  ylab = "Response",
  xlab = "",
  legend = "Legend",
  color = "rainbow",
  fill = "lightblue",
  textsize = 12,
  addmean = TRUE,
  errorbar = TRUE,
  CV = TRUE,
  dec = 3,
  angle = 0,
  posi = "right",
  family = "sans",
  point = "mean_sd",
  sup = NA,
  ylim = NA,
  angle.label = 0
)
}
\arguments{
\item{f1}{Numeric or complex vector with factor 1 levels}

\item{f2}{Numeric or complex vector with factor 2 levels}

\item{response}{Numerical vector containing the response of the experiment.}

\item{norm}{Error normality test (\emph{default} is Shapiro-Wilk)}

\item{homog}{Homogeneity test of variances (\emph{default} is Bartlett)}

\item{mcomp}{Multiple comparison test (Tukey (\emph{default}), LSD, Scott-Knott and Duncan)}

\item{alpha.f}{Level of significance of the F test (\emph{default} is 0.05)}

\item{alpha.t}{Significance level of the multiple comparison test (\emph{default} is 0.05)}

\item{quali}{Defines whether the factor is quantitative or qualitative (\emph{qualitative})}

\item{grau}{Degree of polynomial in case of quantitative factor (\emph{default} is 1)}

\item{transf}{Applies data transformation (default is 1; for log consider 0)}

\item{geom}{Graph type (columns or segments (For simple effect only))}

\item{theme}{ggplot2 theme (\emph{default} is theme_classic())}

\item{ylab}{Variable response name (Accepts the \emph{expression}() function)}

\item{xlab}{Treatments name (Accepts the \emph{expression}() function)}

\item{legend}{Legend title name}

\item{color}{Column chart color (\emph{default} is "rainbow")}

\item{fill}{Defines chart color (to generate different colors for different treatments, define fill = "trat")}

\item{textsize}{Font size}

\item{addmean}{Plot the average value on the graph (\emph{default} is TRUE)}

\item{errorbar}{Plot the standard deviation bar on the graph (In the case of a segment and column graph) - \emph{default} is TRUE}

\item{CV}{Plotting the coefficient of variation and p-value of Anova (\emph{default} is TRUE)}

\item{dec}{Number of cells}

\item{angle}{x-axis scale text rotation}

\item{posi}{Legend position}

\item{family}{Font family}

\item{point}{if quali=F, defines whether to plot all points ("all"), mean ("mean"), standard deviation ("mean_sd") or mean with standard error (\emph{default} - "mean_se").}

\item{sup}{Number of units above the standard deviation or average bar on the graph}

\item{ylim}{y-axis scale}

\item{angle.label}{Label angle}
}
\value{
The table of analysis of variance, the test of normality of errors (Shapiro-Wilk, Lilliefors, Anderson-Darling, Cramer-von Mises, Pearson and Shapiro-Francia), the test of homogeneity of variances (Bartlett or Levene), the test of independence of Durbin-Watson errors, the test of multiple comparisons (Tukey, LSD, Scott-Knott or Duncan) or adjustment of regression models up to grade 3 polynomial, in the case of quantitative treatments. The column chart for qualitative treatments is also returned.
}
\description{
Analysis of an experiment conducted in a completely randomized design in a double factorial scheme using analysis of variance of fixed effects.
}
\note{
The ordering of the graph is according to the sequence in which the factor levels are arranged in the data sheet. The bars of the column and segment graphs are standard deviation.

The function does not perform multiple regression in the case of two quantitative factors.

In the final output when transformation (transf argument) is different from 1, the columns resp and respo in the mean test are returned, indicating transformed and non-transformed mean, respectively.
}
\examples{

#====================================
# Example cloro
#====================================
library(AgroR)
data(cloro)
with(cloro, FAT2DIC(f1, f2, resp, ylab="Number of nodules", legend = "Stages"))

#====================================
# Example corn
#====================================
library(AgroR)
data(corn)
with(corn, FAT2DIC(A, B, Resp, quali=c(TRUE, TRUE),ylab="Heigth (cm)"))
with(corn, FAT2DIC(A, B, Resp, mcomp="sk", quali=c(TRUE, TRUE),ylab="Heigth (cm)"))

}
\references{
Principles and procedures of statistics a biometrical approach Steel & Torry & Dickey. Third Edition 1997

Multiple comparisons theory and methods. Departament of statistics the Ohio State University. USA, 1996. Jason C. Hsu. Chapman Hall/CRC.

Practical Nonparametrics Statistics. W.J. Conover, 1999

Ramalho M.A.P., Ferreira D.F., Oliveira A.C. 2000. Experimentacao em Genetica e Melhoramento de Plantas. Editora UFLA.

Scott R.J., Knott M. 1974. A cluster analysis method for grouping mans in the analysis of variance. Biometrics, 30, 507-512.

Mendiburu, F., & de Mendiburu, M. F. (2019). Package ‘agricolae’. R Package, Version, 1-2.
}
\seealso{
\link{FAT2DIC.art}, \link{FAT2DIC.ad}
}
\author{
Gabriel Danilo Shimizu, \email{shimizu@uel.br}

Leandro Simoes Azeredo Goncalves

Rodrigo Yudi Palhaci Marubayashi
}
\keyword{DIC}
\keyword{Factorial}
