% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibarplot_double.R
\name{ibarplot.double}
\alias{ibarplot.double}
\title{Graph: Invert letters for two factor chart}
\usage{
ibarplot.double(analysis)
}
\arguments{
\item{analysis}{FAT2DIC, FAT2DBC, PSUBDIC or PSUBDBC object}
}
\value{
Return column chart for two factors
}
\description{
invert uppercase and lowercase letters in graph for factorial scheme the subdivided plot with significant interaction
}
\examples{
data(covercrops)
attach(covercrops)
a=FAT2DBC(A, B, Bloco, Resp, ylab=expression("Yield"~(Kg~"100 m"^2)),
legend = "Cover crops",alpha.f = 0.3,family = "serif")
ibarplot.double(a)
}
