% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laranja.R
\docType{data}
\name{laranja}
\alias{laranja}
\title{Dataset: Orange plants under different rootstocks}
\format{
data.frame containing data set
  \describe{
  \item{\code{f1}}{Categorical vector with treatments}
  \item{\code{bloco}}{Categorical vector with block}
  \item{\code{resp}}{Numeric vector with number of fruits per plant}
  }
}
\usage{
data(laranja)
}
\description{
An experiment was conducted with the objective of studying the behavior
of nine rootstocks for the Valencia orange tree. The data set refers to
the 1973 evaluation (12 years old). The rootstocks are: T1: Tangerine Sunki;
T2: National rough lemon; T3: Florida rough lemon; T4: Cleopatra tangerine;
T5: Citranger-troyer; T6: Trifoliata; T7: Clove Tangerine; T8: Country orange;
T9: Clove Lemon. The number of fruits per plant was evaluated.
}
\examples{
data(laranja)
}
\references{
Planejamento e Analise Estatistica de Experimentos Agronomicos (2013) - Decio Barbin - pg. 72
}
\seealso{
\link{cloro}, \link{enxofre}, \link{mirtilo}, \link{pomegranate}, \link{porco}, \link{sensorial}, \link{simulate1}, \link{simulate2}, \link{simulate3}, \link{tomate}, \link{weather}, \link{phao}, \link{passiflora}, \link{aristolochia}
}
\keyword{datasets}
