% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirtilo.R
\docType{data}
\name{mirtilo}
\alias{mirtilo}
\title{Dataset: Cutting blueberry data}
\format{
data.frame containing data set
  \describe{
  \item{\code{trat}}{Categorical vector with treatments}
  \item{\code{exp}}{Categorical vector with experiment}
  \item{\code{bloco}}{Categorical vector with block}
  \item{\code{resp}}{Numeric vector}
  }
}
\usage{
data(mirtilo)
}
\description{
An experiment was carried out in order to evaluate the rooting
(resp1) of blueberry cuttings as a function of the cutting size
(Treatment Colume). This experiment was repeated three times
(Location column) and a randomized block design with four
replications was adopted.
}
\examples{
data(mirtilo)
attach(mirtilo)
}
\seealso{
\link{cloro}, \link{enxofre}, \link{laranja}, \link{pomegranate}, \link{porco}, \link{sensorial}, \link{simulate1}, \link{simulate2}, \link{simulate3}, \link{tomate}, \link{weather}
}
\keyword{datasets}
