% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepper_dataset.R
\docType{data}
\name{pepper}
\alias{pepper}
\title{Dataset: Pepper}
\format{
data.frame containing data set
  \describe{
  \item{\code{Acesso}}{Categorical vector with accessions}
  \item{\code{MS}}{Numeric vector com dry mass}
  \item{\code{VitC}}{Numeric vector with Vitamin C}
  }
}
\usage{
data(pepper)
}
\description{
A vegetable breeder is characterizing five mini pepper
accessions from the State University of Londrina germplasm bank for
agronomic and biochemical variables. The experiment was conducted in
a completely randomized design with four replications
}
\examples{
data(pepper)
}
\seealso{
\link{enxofre}, \link{laranja}, \link{mirtilo}, \link{pomegranate}, \link{porco}, \link{sensorial}, \link{simulate1}, \link{simulate2}, \link{simulate3}, \link{tomate}, \link{weather}, \link{phao}, \link{passiflora}, \link{aristolochia}
}
\keyword{datasets}
