% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phao.R
\docType{data}
\name{phao}
\alias{phao}
\title{Dataset: Osmocote in \emph{Phalaenopsis} sp.}
\format{
data.frame containing data set
  \describe{
  \item{\code{dose}}{Numeric vector with doses}
  \item{\code{comp}}{Numeric vector with leaf length}
  }
}
\usage{
data(phao)
}
\description{
The objective of the work was to evaluate the
effect of doses of osmocote (15-09-12-N-P2O5-K2O,
respectively) on the initial development of the
orchid \emph{Phalaenopsis} sp. The osmocote fertilizer
was added in the following doses: 0, 2, 4, 6 and
8 g vase-1. After twelve months, leaf length was
evaluated.
}
\examples{
data(phao)
}
\references{
de Paula, J. C. B., Junior, W. A. R., Shimizu, G. D., Men, G. B., & de Faria, R. T. (2020). Fertilizante de liberacao controlada no crescimento inicial da orquidea \emph{Phalaenopsis} sp. Revista Cultura Agronomica, 29(2), 289-299.
}
\seealso{
\link{pomegranate}, \link{passiflora}, \link{cloro}, \link{enxofre}, \link{laranja}, \link{mirtilo}, \link{porco}, \link{sensorial}, \link{simulate1}, \link{simulate2}, \link{simulate3}, \link{tomate}, \link{weather}
}
\keyword{datasets}
