% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interaction_function.R
\name{plot_interaction}
\alias{plot_interaction}
\title{Graph: Interaction plot}
\usage{
plot_interaction(a)
}
\arguments{
\item{a}{FAT2DIC, FAT2DBC, PSUBDIC or PSUBDBC object}
}
\value{
Returns an interaction graph with averages and letters from the multiple comparison test
}
\description{
Performs an interaction graph from an output of the FAT2DIC, FAT2DBC, PSUBDIC or PSUBDBC commands.
}
\examples{
data(cloro)
attach(cloro)
a=FAT2DIC(f1, f2, resp)
plot_interaction(a)
}
\author{
Gabriel Danilo Shimizu, \email{shimizu@uel.br}

Leandro Simoes Azeredo Goncalves

Rodrigo Yudi Palhaci Marubayashi
}
