% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotjitter_function.R
\name{plot_jitter}
\alias{plot_jitter}
\title{Graph: Column, box or segment chart with observations}
\usage{
plot_jitter(model)
}
\arguments{
\item{model}{DIC, DBC or DQL object}
}
\value{
Returns with graph of boxes, columns or segments with all the observations represented in the graph.
}
\description{
The function performs the construction of graphs of boxes, columns or segments with all the observations represented in the graph.
}
\examples{
data("pomegranate")
a=with(pomegranate,DIC(trat,WL,geom="point"))
plot_jitter(a)
}
\author{
Gabriel Danilo Shimizu, \email{shimizu@uel.br}

Leandro Simoes Azeredo Goncalves

Rodrigo Yudi Palhaci Marubayashi
}
