% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_function.R
\name{polynomial}
\alias{polynomial}
\title{Analysis: Linear regression graph}
\usage{
polynomial(
  trat,
  resp,
  ylab = "Response",
  xlab = "Independent",
  grau = NA,
  theme = theme_bw(),
  color = "gray80",
  posi = "top",
  textsize = 12,
  se = FALSE,
  ylim = NA,
  family = "sans",
  pointsize = 4.5,
  decimal = "."
)
}
\arguments{
\item{trat}{Numerical vector with treatments (Declare as numeric)}

\item{resp}{Numerical vector containing the response of the experiment.}

\item{ylab}{Dependent variable name (Accepts the \emph{expression}() function)}

\item{xlab}{Independent variable name (Accepts the \emph{expression}() function)}

\item{grau}{Degree of the polynomial (1,2 or 3)}

\item{theme}{ggplot2 theme (\emph{default} is theme_bw())}

\item{color}{Graph color (\emph{default} is gray80)}

\item{posi}{Legend position}

\item{textsize}{Font size}

\item{se}{Adds confidence interval (\emph{default} is FALSE)}

\item{ylim}{y-axis scale}

\item{family}{Font family}

\item{pointsize}{Point size}

\item{decimal}{Decimal separate}
}
\value{
Returns linear, quadratic or cubic regression analysis.
}
\description{
Linear regression analysis of an experiment with a quantitative factor or isolated effect of a quantitative factor
}
\examples{
data("phao")
attach(phao)
polynomial(dose,comp, grau = 2)
}
\seealso{
\link{polynomial2}
}
\author{
Gabriel Danilo Shimizu, \email{shimizu@uel.br}

Leandro Simoes Azeredo Goncalves

Rodrigo Yudi Palhaci Marubayashi
}
\keyword{Experimental}
\keyword{Regression}
