% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radargraph_function.R
\name{radargraph}
\alias{radargraph}
\title{Graph: Circular column chart}
\usage{
radargraph(model, ylim = NA, labelsize = 4, transf = FALSE)
}
\arguments{
\item{model}{DIC, DBC or DQL object}

\item{ylim}{y-axis limit}

\item{labelsize}{Font size of the labels}

\item{transf}{If the data has been transformed (\emph{default} is FALSE)}
}
\value{
Returns pie chart with averages and letters from the Scott-Knott cluster test
}
\description{
Circular column chart of an experiment with a factor of interest or isolated effect of a factor
}
\examples{
data("laranja")
a=with(laranja, DBC(trat,bloco,resp, mcomp = "sk"))
radargraph(a)
}
\seealso{
\link{barplot_positive}, \link{sk_graph}, \link{plot_TH}, \link{corgraph}, \link{spider_graph}, \link{line_plot}
}
\author{
Gabriel Danilo Shimizu, \email{shimizu@uel.br}

Leandro Simoes Azeredo Goncalves

Rodrigo Yudi Palhaci Marubayashi
}
