% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seggraph_function.R
\name{seg_graph}
\alias{seg_graph}
\title{Graph: Segment graph for one factor}
\usage{
seg_graph(model, fill = "lightblue", horiz = TRUE, pointsize = 4.5)
}
\arguments{
\item{model}{DIC, DBC or DQL object}

\item{fill}{fill bars}

\item{horiz}{Horizontal Column (\emph{default} is TRUE)}

\item{pointsize}{Point size}
}
\value{
Returns a segment chart for one factor
}
\description{
This is a function of the bar graph for one factor
}
\examples{
data("laranja")
a=with(laranja, DBC(trat, bloco, resp,
     mcomp = "sk",angle=45,
     ylab = "Number of fruits/plants"))
seg_graph(a,horiz = FALSE)
}
\seealso{
\link{radargraph}, \link{barplot_positive}, \link{plot_TH}, \link{corgraph}, \link{spider_graph}, \link{line_plot}
}
\author{
Gabriel Danilo Shimizu, \email{shimizu@uel.br}

Leandro Simoes Azeredo Goncalves

Rodrigo Yudi Palhaci Marubayashi
}
