% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seggraph2_function.R
\name{seg_graph2}
\alias{seg_graph2}
\title{Graph: Point graph for one factor model 2}
\usage{
seg_graph2(
  model,
  theme = theme_gray(),
  pointsize = 4,
  pointshape = 16,
  horiz = TRUE,
  vjust = -0.6
)
}
\arguments{
\item{model}{DIC, DBC or DQL object}

\item{theme}{ggplot2 theme}

\item{pointsize}{Point size}

\item{pointshape}{Format point (default is 16)}

\item{horiz}{Horizontal Column (\emph{default} is TRUE)}

\item{vjust}{vertical adjusted}
}
\value{
Returns a point chart for one factor
}
\description{
This is a function of the point graph for one factor
}
\examples{
data("laranja")
a=with(laranja, DBC(trat, bloco, resp,
       mcomp = "sk",angle=45,
       ylab = "Number of fruits/plants"))
seg_graph2(a,horiz = FALSE)
}
\seealso{
\link{radargraph}, \link{barplot_positive}, \link{plot_TH}, \link{corgraph}, \link{spider_graph}, \link{line_plot}
}
\author{
Gabriel Danilo Shimizu, \email{shimizu@uel.br}

Leandro Simoes Azeredo Goncalves

Rodrigo Yudi Palhaci Marubayashi
}
