% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate1.R
\docType{data}
\name{simulate1}
\alias{simulate1}
\title{Dataset: Simulated data dict}
\format{
data.frame containing data set
  \describe{
  \item{\code{tempo}}{Categorical vector with time}
  \item{\code{trat}}{Categorical vector with treatment}
  \item{\code{resp}}{Categorical vector with response}
  }
}
\usage{
data(simulate1)
}
\description{
Simulated data from a completely randomized experiment with multiple assessments over time
}
\examples{
data(simulate1)
}
\seealso{
\link{cloro}, \link{enxofre}, \link{laranja}, \link{mirtilo}, \link{pomegranate}, \link{porco}, \link{sensorial}, \link{simulate2}, \link{simulate3}, \link{tomate}, \link{weather}, \link{phao}, \link{passiflora}, \link{aristolochia}
}
\keyword{datasets}
