% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_graph_function.R
\name{sk_graph}
\alias{sk_graph}
\title{Graph: Scott-Knott graphics}
\usage{
sk_graph(model, dec = 3, horiz = TRUE, transf = FALSE)
}
\arguments{
\item{model}{DIC, DBC or DQL object}

\item{dec}{Number of cells}

\item{horiz}{Horizontal Column (\emph{default} is TRUE)}

\item{transf}{If the data has been transformed (\emph{default} is FALSE)}
}
\value{
Returns a bar chart with columns separated by color according to the Scott-Knott test
}
\description{
This is a function of the bar graph for the Scott-Knott test
}
\examples{
data("laranja")
attach(laranja)
a=DBC(trat, bloco, resp,
     mcomp = "sk",angle=45,
     ylab = "Number of fruits/plants")
sk_graph(a,dec = 0,horiz = FALSE)
}
\seealso{
\link{radargraph}, \link{barplot_positive}, \link{plot_TH}, \link{corgraph}, \link{spider_graph}, \link{line_plot}
}
\author{
Gabriel Danilo Shimizu, \email{shimizu@uel.br}

Leandro Simoes Azeredo Goncalves

Rodrigo Yudi Palhaci Marubayashi
}
