% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croqui_function.R
\name{sketch}
\alias{sketch}
\title{Utils: Experimental sketch}
\usage{
sketch(trat, trat1 = NULL, r, design = "dic", pos = "line", ncol = NA)
}
\arguments{
\item{trat}{Vector with factor A levels}

\item{trat1}{Vector with levels of factor B (Set to NULL if not factorial or psub)}

\item{r}{Number of repetitions}

\item{design}{Experimental design ("dic", "dbc", "dql","psubdic","psubdbc","fat2dic","fat2dbc")}

\item{pos}{Repeat position (line or column)}

\item{ncol}{Default is NA. Warning!!! Use only in a completely randomized design}
}
\value{
Returns an experimental sketch according to the specified design.
}
\description{
Experimental sketching function
}
\note{
The sketches have only a rectangular shape, and the blocks (in the case of randomized blocks) can be in line or in a column.
Mendiburu, F., & de Mendiburu, M. F. (2019). Package ‘agricolae’. R Package, Version, 1-2.
}
\examples{
Trat=paste("Treatments",1:6)

#=============================
# Completely randomized design
#=============================
sketch(Trat,r=3)

#=============================
# Randomized block design
#=============================
sketch(Trat, r=3, design="dbc")

#=============================
# Completely randomized experiments in double factorial
#=============================
sketch(trat=c(1,2),
       trat1=c(1,2,3),
       design = "fat2dic",
       r=3)
}
\author{
Gabriel Danilo Shimizu, \email{shimizu@uel.br}

Leandro Simoes Azeredo Goncalves

Rodrigo Yudi Palhaci Marubayashi
}
\keyword{croqui}
\keyword{experimental}
