% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soybean_dataset.R
\docType{data}
\name{soybean}
\alias{soybean}
\title{Dataset: Soybean}
\format{
data.frame containing data set
  \describe{
  \item{\code{cult}}{numeric vector with treatment}
  \item{\code{bloc}}{numeric vector with block}
  \item{\code{prod}}{Numeric vector with grain yield}
  }
}
\usage{
data("soybean")
}
\description{
An experiment was carried out to evaluate the grain yield (kg ha-1) of ten different
commercial soybean cultivars in the municipality of Londrina/Parana. The experiment was carried out in the design of
randomized complete blocks with four replicates per treatment.
}
\examples{
data(soybean)
}
\seealso{
\link{cloro}, \link{laranja}, \link{enxofre}, \link{laranja}, \link{mirtilo}, \link{passiflora}, \link{phao}, \link{porco}, \link{pomegranate}, \link{simulate1}, \link{simulate2}, \link{simulate3}, \link{tomate}, \link{weather}
}
\keyword{datasets}
