% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabledesc_function.R
\name{tabledesc}
\alias{tabledesc}
\title{Descriptive: Table descritive analysis}
\usage{
tabledesc(data, fun = mean)
}
\arguments{
\item{data}{data.frame containing the first column with the categorical variable and the remaining response columns}

\item{fun}{Function of descriptive statistics (default is mean)}
}
\value{
Returns a data.frame with a measure of dispersion or position from a dataset and separated by a factor
}
\description{
Function for generating a data.frame with averages or other descriptive measures grouped by a categorical variable
}
\examples{
data(pomegranate)
tabledesc(pomegranate)
}
\author{
Gabriel Danilo Shimizu, \email{shimizu@uel.br}

Leandro Simoes Azeredo Goncalves

Rodrigo Yudi Palhaci Marubayashi
}
\keyword{descriptive}
