% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbarplotreverse_function.R
\name{TBARPLOT.reverse}
\alias{TBARPLOT.reverse}
\title{Graph: Reverse graph of DICT, DBCT and DQL output when geom="bar"}
\usage{
TBARPLOT.reverse(plot.t)
}
\arguments{
\item{plot.t}{DICT, DBCT or DQLT output when geom="bar"}
}
\value{
Returns a reverse graph of the output of DICT, DBCT or DQLT when geom="bar".
}
\description{
The function performs the construction of a reverse graph on the output of DICT, DBCT and DQL when geom="bar".
}
\note{
All layout and subtitles are imported from DICT, DBCT and DQLT functions
}
\examples{
data(simulate1)
a=with(simulate1, DICT(trat, tempo, resp,geom="bar",sup=40))
TBARPLOT.reverse(a)
}
\seealso{
\link{DICT}, \link{DBCT}, \link{DQLT}
}
\author{
Gabriel Danilo Shimizu, \email{shimizu@uel.br}

Leandro Simoes Azeredo Goncalves

Rodrigo Yudi Palhaci Marubayashi
}
\keyword{Experimental}
