% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomate.R
\docType{data}
\name{tomate}
\alias{tomate}
\title{Dataset: Tomato data}
\format{
data.frame containing data set
  \describe{
  \item{\code{parc}}{Categorical vector with plot}
  \item{\code{subp}}{Categorical vector with split-plot}
  \item{\code{bloco}}{Categorical vector with block}
  \item{\code{resp}}{Numeric vector}
  }
}
\usage{
data(tomate)
}
\description{
An experiment conducted in a randomized block design in a split
plot scheme was developed in order to evaluate the efficiency of
bacterial isolates in the development of tomato cultivars. The
experiment counted a total of 24 trays; each block (in a total
of four blocks), composed of 6 trays, in which each tray contained
a treatment (6 isolates). Each tray was seeded with 4 different
genotypes, each genotype occupying 28 cells per tray. The trays
were randomized inside each block and the genotypes were randomized
inside each tray.
}
\examples{
data(tomate)
}
\seealso{
\link{cloro}, \link{enxofre}, \link{laranja}, \link{mirtilo}, \link{pomegranate}, \link{porco}, \link{sensorial}, \link{simulate1}, \link{simulate2}, \link{simulate3}, \link{weather}, \link{aristolochia}, \link{phao}, \link{passiflora}
}
\keyword{datasets}
