% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\docType{data}
\name{weather}
\alias{weather}
\title{Dataset: Weather data}
\format{
data.frame containing data set
  \describe{
  \item{\code{Data}}{POSIXct vector with dates}
  \item{\code{tempo}}{Numeric vector with time}
  \item{\code{Tmax}}{Numeric vector with maximum temperature}
  \item{\code{Tmed}}{Numeric vector with mean temperature}
  \item{\code{Tmin}}{Numeric vector with minimum temperature}
  \item{\code{UR}}{Numeric vector with relative humidity}
  }
}
\usage{
data(weather)
}
\description{
Climatic data from 01 November 2019 to 30 June 2020 in the municipality of Londrina-PR, Brazil. Data from the Instituto de Desenvolvimento Rural do Parana (IDR-PR)
}
\examples{
data(weather)
}
\seealso{
\link{cloro}, \link{enxofre}, \link{laranja}, \link{mirtilo}, \link{pomegranate}, \link{porco}, \link{sensorial}, \link{simulate1}, \link{simulate2}, \link{simulate3}, \link{tomate}, \link{aristolochia}, \link{phao}, \link{passiflora}
}
\keyword{datasets}
