% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbc_function.R
\name{DBC}
\alias{DBC}
\title{Analysis: Randomized block design}
\usage{
DBC(
  trat,
  block,
  response,
  norm = "sw",
  homog = "bt",
  mcomp = "tukey",
  quali = TRUE,
  alpha.f = 0.05,
  alpha.t = 0.05,
  transf = 1,
  test = "parametric",
  grau = 1,
  geom = "bar",
  theme = theme_classic(),
  sup = NA,
  CV = TRUE,
  ylab = "response",
  xlab = "",
  textsize = 12,
  fill = "lightblue",
  angle = 0,
  family = "sans",
  dec = 3,
  addmean = TRUE,
  errorbar = TRUE,
  posi = "top",
  point = "mean_sd",
  angle.label = 0
)
}
\arguments{
\item{trat}{Numerical or complex vector with treatments}

\item{block}{Numerical or complex vector with blocks}

\item{response}{Numerical vector containing the response of the experiment.}

\item{norm}{Error normality test (\emph{default} is Shapiro-Wilk)}

\item{homog}{Homogeneity test of variances (\emph{default} is Bartlett)}

\item{mcomp}{Multiple comparison test (Tukey (\emph{default}), LSD, Scott-Knott and Duncan)}

\item{quali}{Defines whether the factor is quantitative or qualitative (\emph{default} is qualitative)}

\item{alpha.f}{Level of significance of the F test (\emph{default} is 0.05)}

\item{alpha.t}{Significance level of the multiple comparison test (\emph{default} is 0.05)}

\item{transf}{Applies data transformation (default is 1; for log consider 0)}

\item{test}{"parametric" - Parametric test or "noparametric" - non-parametric test}

\item{grau}{Degree of polynomial in case of quantitative factor (\emph{default} is 1)}

\item{geom}{graph type (columns, boxes or segments)}

\item{theme}{ggplot2 theme (\emph{default} is theme_classic())}

\item{sup}{Number of units above the standard deviation or average bar on the graph}

\item{CV}{Plotting the coefficient of variation and p-value of Anova (\emph{default} is TRUE)}

\item{ylab}{Variable response name (Accepts the \emph{expression}() function)}

\item{xlab}{Treatments name (Accepts the \emph{expression}() function)}

\item{textsize}{Font size}

\item{fill}{Defines chart color (to generate different colors for different treatments, define fill = "trat")}

\item{angle}{x-axis scale text rotation}

\item{family}{Font family}

\item{dec}{Number of cells}

\item{addmean}{Plot the average value on the graph (\emph{default} is TRUE)}

\item{errorbar}{Plot the standard deviation bar on the graph (In the case of a segment and column graph) - \emph{default} is TRUE}

\item{posi}{Legend position}

\item{point}{Defines whether to plot mean ("mean"), mean with standard deviation ("mean_sd" - \emph{default}) or mean with standard error (\emph{default} - "mean_se").}

\item{angle.label}{label angle}
}
\value{
The table of analysis of variance, the test of normality of errors (Shapiro-Wilk, Lilliefors, Anderson-Darling, Cramer-von Mises, Pearson and Shapiro-Francia), the test of homogeneity of variances (Bartlett or Levene), the test of independence of Durbin-Watson errors, the test of multiple comparisons (Tukey, LSD, Scott-Knott or Duncan) or adjustment of regression models up to grade 3 polynomial, in the case of quantitative treatments. Non-parametric analysis can be used by the Friedman test. The column, segment or box chart for qualitative treatments is also returned. The function also returns a standardized residual plot.
}
\description{
This is a function of the AgroR package for statistical analysis of experiments conducted in a randomized block and balanced design with a factor considering the fixed model.
}
\note{
The ordering of the graph is according to the sequence in which the factor levels are arranged in the data sheet. The bars of the column and segment graphs are standard deviation.

CV and p-value of the graph indicate coefficient of variation and p-value of the F test of the analysis of variance.

In the final output when transformation (transf argument) is different from 1, the columns resp and respo in the mean test are returned, indicating transformed and non-transformed mean, respectively.
}
\examples{
library(AgroR)

#=============================
# Example laranja
#=============================
data(laranja)
attach(laranja)
DBC(trat, bloco, resp, mcomp = "sk", angle=45, ylab = "Number of fruits/plants")

#=============================
# Friedman test
#=============================
DBC(trat, bloco, resp, test="noparametric", ylab = "Number of fruits/plants")

#=============================
# Example soybean
#=============================
data(soybean)
with(soybean, DBC(cult, bloc, prod,
                  ylab=expression("Grain yield"~(kg~ha^-1))))
}
\references{
Principles and procedures of statistics a biometrical approach Steel, Torry and Dickey. Third Edition 1997

Multiple comparisons theory and methods. Departament of statistics the Ohio State University. USA, 1996. Jason C. Hsu. Chapman Hall/CRC.

Practical Nonparametrics Statistics. W.J. Conover, 1999

Ramalho M.A.P., Ferreira D.F., Oliveira A.C. 2000. Experimentacao em Genetica e Melhoramento de Plantas. Editora UFLA.

Scott R.J., Knott M. 1974. A cluster analysis method for grouping mans in the analysis of variance. Biometrics, 30, 507-512.

Mendiburu, F., and de Mendiburu, M. F. (2019). Package ‘agricolae’. R Package, Version, 1-2.
}
\seealso{
\link{DIC}, \link{DQL}
}
\author{
Gabriel Danilo Shimizu, \email{shimizu@uel.br}

Leandro Simoes Azeredo Goncalves

Rodrigo Yudi Palhaci Marubayashi
}
\keyword{DBC}
\keyword{Experimental}
